app/requirements.txt
Django
supervisor 
gunicorn 
djangorestframework 
simplejson
django-debug-toolbar 
python-dateutil 
reportlab 
django-ipware
boto3 
requests 
djangorestframework-simplejwt 
django-cors-headers
django-filter
django-globals 
psycopg2 
psycopg2-binary 
postgres 
drf-yasg